requirements.txt
# Web Framework
Flask>=3.0
Werkzeug>=3.0

# AI & LLM Integration
google-generativeai>=0.4
openai>=1.12

# API Clients & Data Handling
requests>=2.31
python-dotenv>=1.0
pandas>=2.2
beautifulsoup4>=4.12

# Scheduling
APScheduler>=3.10

# Security & Search Tools
serpapi>=0.1.5
ipstack>=0.1.0

# Production (optional): use for production deploy instead of python app.py
# gunicorn>=23.0
